+++ pkgs/jnigen/CHANGELOG.md
@@ -12,6 +12,8 @@
 - Fixed a [bug](https://github.com/dart-lang/native/issues/2250) where classes
   that inherited a generic without specifying all type parameters were not
   generated.
+- Added the ability to generate Kotlin stdlib classes without providing the
+  class path.
 
 ## 0.14.1